requirements.txt
Django==4.2.6
django-cors-headers==3.13.0
django-filter==22.1
django_jazzmin==2.6.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
factory-boy==3.3.0
geopy==2.4.1
Pillow==9.2.0
psycopg2-binary==2.9.9
pytest-django==4.8.0
python-dotenv==1.0.0
pytz==2022.4